Split off patterns.
